refactor(switchdash): split the sidebar into an agent tree and a room tree (CHOO-1875)
The two groupings were one component with a branch at the top, which made them
look like the same tree with its levels swapped. They are not. Grouped by
agent, a room is a property of a session — a heading over some of an agent's
work, absent when nothing is running. Grouped by room, a room is the subject:
listed because it exists, with rows that act on the room itself.

Forcing them together had already cost a bug — the room tree was written by
analogy and quietly dropped the agent-expand check, so an agent's chevron said
collapsed while its sessions stayed on screen.

Split them into their own components, each free to render what its view means,
over shared data helpers that answer only "which agents are in scope" and
"which sessions are theirs". No behaviour change.
